[os] Move structs into own header file
